refactor : 사용자, 좋아요 엔티티에 baseEntity 상속 추가
